+++ cpp/server/foobar2000/plugin.hpp
@@ -7,8 +7,7 @@
 #include "../settings.hpp"
 #include "../server_host.hpp"
 
-namespace msrv {
-namespace player_foobar2000 {
+namespace msrv::player_foobar2000 {
 
 class Plugin
 {
@@ -23,6 +22,8 @@ public:
 
     void reconfigure();
 
+    static Path getProfileDir();
+
 private:
     static Plugin* current_;
 
@@ -32,5 +33,4 @@ private:
     MSRV_NO_COPY_AND_ASSIGN(Plugin);
 };
 
-}
 }